Added simpler helix example.
Moved pticklabel to LeftTicks, etc., as the argument plabel. Added pTick and
ptick pen types for drawing big and small ticks with different sizes,
bool beginlabel and endlabel for suppressing first and last labels,
and extend for drawing ticks across the graph (useful for superimposing a grid
on the graph).

Improved optional first and last tick/label suppression.


git-svn-id: svn://svn.code.sf.net/p/asymptote/code/trunk@880 fb5b39b3-7a14-0410-b0a0-e9e911d1d576