Added interface to GNU readline library to allow editing with history when reading data from stdin. Updated getstring, getreal, etc. in strings.asy to use this new readline function.
Added complement(int[] a, T[] b) function to return the complement of the
integer array a in {1,2,...,b.length}, so that b[complement(a,b)] yields the
complement of b[a].

Generated dataSettings from a templated struct; added intSetting.

Added historylines option (default is still 1000).

Added array check to arrayConditional.

Updated documentation.


git-svn-id: svn://svn.code.sf.net/p/asymptote/code/trunk@1293 fb5b39b3-7a14-0410-b0a0-e9e911d1d576